fix: run compiler in task
This lets the compiler stdout messgaes to be logged to the client while
the RPC is still happening.

Previously, the port would message back to the process that was
synchronously waiting on the RPC to finish, so it didn't send any log
messages to the LSP process until the compiler was done.

This made it look like the initial compiler was hanging.

Fixes #60