fix: swap out dets for sqlite3
Also fixes a bug with go to definition. it turns out that multiple
references were being stored that overlapped in range, so we'd get
multiple references, which didn't match the pattern, so it returned nil.

Now, sqlite easily deletes any references that overlap with the one we
are inserting in that moment.