fix: use first reference
when the compiler is tracing macros, the actual source code gets
expanded first and that is what we want, later, it will trace the
expanded code, and it will exist in ranges that overlap with the actual
source code.

I believe we can rely on the integer id being lower for the first one,
and use an order by clause.