chore(nix): improve flake and build in CI (#312)
* style(nix): centralise the specification of Erlang and Elixir versions

* chore(nix): switch to nixos-unstable channel

The master branch is where PRs are merged into. Binary cache is more likely
available on stable and unstable branches, so switch to unstable to use latest
packages.

* chore(nix): downgrade to Erlang R25 to prevent rebuild of rebar3

In NixOS/nixpkgs, the latest version of BEAM packages may not have all packages
built on Hydra. In fact, rebar3 is not available from binary cache, which
requires the package user to rebuild the package. Downgrading to the previous
version of Erlang VM solves this issue.

* fix(nix): fix hash mismatch of mixFodDeps

* fix(nix)!: remove the broken flake app

The path points to a location which apparently no longer exists.

* chore(nix): add meta data of the Nix package

Setting meta.mainProgram is recommended now so `nix run` consistently works.

* style(nix): Use a map to manage the platform mappings

* style(nix): tidy preInstall and postInstall hooks

* style(nix): tidy

* feat(nix): make the Erlang and Elixir versions overridable

The user may want to use the same versions as his/her project under development,
so it makes sense to allow overriding of beamPackages and elixir.

* fix(nix): fix installation on ARM64 Linux

* fix(nix): Revert to Erlang R26 as it is a requirement

* style(nix): apply alejandra formatter

* ci: add nix builds

* fix(nix): Use the same version of Elixir for fetchMixDeps

* update hash

* fix ci

---------

Co-authored-by: Mitchell Hanberg <mitch@mitchellhanberg.com>