feat(runtime): Support workspace shims with `elixir_exec` option
Relates to #334

IDEs like to work with shims rather than tools that alter path
(such as `mise`/`rtx`). As a result, one must launch their IDE
manually from a terminal session with all the `PATH` loaded in
order to properly find the `elixir` executable. This works fine,
but switching between projects/workspaces that have varying
Elixir/OTP versions will most likely be using the version loaded
when launching the program from terminal.

One resolve might be to include shims in the `PATH` globally.
However, that negates some of the reasoning to move to alternate
tools like `mise.

Instead, this allows setting the `elixir_exec` option to point
directly at a shim. When used in workspace directories, that
should also be able to load the correct versions defined in the
tool version files. This also allows keeping the need for a shim
isolated to the LS tool rather than global use