fix(completions): correctly insert cursor position
The cursor was being inserted in a strange matter, which made the
modified ast not have a `__cursor__()` function, so we couldn't find the
cursor.

Really, all we need to do is insert the cursor on its own line between
the line.

For example, if you were to put the cursor into the middle of this line
of code

```elixir
Enum.map(some_list, fn row ->   end)
```

It would insert the cursor like so

```elixir
Enum.map(some_list, fn row ->
__cursor__()
end)
```

Fixes #406